+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Adds LineViewHelper
 - Adds speaking settings
 - Adds partial overwriting of text padding, [#25](https://github.com/bithost-gmbh/pdfviewhelpers/issues/25)
+- Adds width and padding to ColumnViewHelper
 - Replaces EmptyFPDI and EmptyTCPDF by BasePDF
 - Removes example class BithostTCPDF
 - Fixes MultiColumnViewHelper columns not correctly initialized